+++ CHANGELOG.md
@@ -620,3 +620,5 @@ Developmental progress by date is stored in this file.
 - Make some recent comments more descriptive
 - pwa_install only supports Chromium browsers, so added a fallback that leads to a different screen that will show users how to manually install as a PWA (pwa_install_guide.dart is the tutorial screen)
 - Made the PWA install button invisible when the app is already installed
+- Added Back button tooltip to Explore tab to be consistent with other pages
+- Added tooltip explaining what a PWA is in the Install PWA App button